Demo of how to create VCluster and install Jaeger
This is a refactor of `vcluster` branch from the forked repo.

The key here is that the VCluster promise now registers itself
to Kratix allowing any future resource requests to be scheduled.
The labels used for scheduling are quite generic (`environment: dev`)
and also down to specific (`vclusterName: xxx`). The specific name
label is used to make a resource request for Jaeger.

Then there is a new Jaeger promise which exposes `clusterSelector` in
the xaasCrd allowing the Resource Request from the VCluster promise
to schedule to the specificly created cluster.